Update, fix and reorganize the integration tests
