+++ miscellaneous/roboconf-integration-tests/pom.xml
@@ -253,6 +253,12 @@
 			<version>${jersey.stack.version}</version>
 			<scope>test</scope>
 		</dependency>
+		
+		<dependency>
+			<groupId>com.sun.jersey.contribs</groupId>
+			<artifactId>jersey-multipart</artifactId>
+			<scope>test</scope>
+		</dependency>
 	</dependencies>
 	
 	<build>
