+++ miscellaneous/roboconf-integration-tests/src/test/java/net/roboconf/integration/probes/DmTest.java
@@ -25,29 +25,20 @@
 
 package net.roboconf.integration.probes;
 
-import java.util.List;
+import net.roboconf.integration.tests.internal.ItUtils;
 
 import org.ops4j.pax.exam.Configuration;
 import org.ops4j.pax.exam.Option;
 
 /**
+ * A base class to run the DM's distribution.
  * @author Vincent Zurczak - Linagora
  */
-public abstract class DmTest extends AbstractTest {
-
-	@Override
-	protected String getArtifactId() {
-		return "roboconf-karaf-dist-dm";
-	}
-
-	@Override
-	protected String getDirectorySuffix() {
-		return "dm";
-	}
+public abstract class DmTest {
 
 	@Configuration
 	public Option[] config() throws Exception {
-		List<Option> options = getBaseOptions();
-		return options.toArray( new Option[ options.size()]);
+		ItConfigurationBean bean = new ItConfigurationBean( "roboconf-karaf-dist-dm", "dm" );
+		return ItUtils.getBaseOptions( bean );
 	}
 }
