+++ miscellaneous/roboconf-integration-tests/src/test/java/net/roboconf/integration/tests/paxrunner/AgentInitializationTest.java
@@ -23,12 +23,14 @@
  * limitations under the License.
  */
 
-package net.roboconf.integration.tests;
+package net.roboconf.integration.tests.paxrunner;
 
 import static org.ops4j.pax.exam.CoreOptions.mavenBundle;
 import static org.ops4j.pax.exam.CoreOptions.systemProperty;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import javax.inject.Inject;
@@ -37,7 +39,6 @@ import net.roboconf.agent.internal.Agent;
 import net.roboconf.agent.internal.AgentMessageProcessor;
 import net.roboconf.agent.internal.misc.HeartbeatTask;
 import net.roboconf.agent.internal.misc.PluginMock;
-import net.roboconf.core.Constants;
 import net.roboconf.core.internal.tests.TestUtils;
 import net.roboconf.core.model.beans.ApplicationTemplate;
 import net.roboconf.core.model.beans.Instance;
@@ -45,8 +46,8 @@ import net.roboconf.core.model.beans.Instance.InstanceStatus;
 import net.roboconf.core.model.helpers.InstanceHelpers;
 import net.roboconf.dm.management.ManagedApplication;
 import net.roboconf.dm.management.Manager;
-import net.roboconf.integration.probes.AbstractTest;
 import net.roboconf.integration.probes.DmTest;
+import net.roboconf.integration.tests.internal.ItUtils;
 import net.roboconf.integration.tests.internal.MyHandler;
 import net.roboconf.integration.tests.internal.MyTargetResolver;
 import net.roboconf.integration.tests.internal.RoboconfPaxRunner;
@@ -65,8 +66,8 @@ import org.ops4j.pax.exam.spi.reactors.PerMethod;
  * A set of tests for the agent's initialization.
  * <p>
  * We launch a Karaf installation with an agent in-memory. We load
- * an application and instantiates a scoped instance which is not a root.
- * The new agent must send an initial message to the DM to indicate it is alive.
+ * an application and instantiates a root instance. The new agent
+ * must send an initial message to the DM to indicate it is alive.
  * It must then receive its model from the DM.
  * </p>
  *
@@ -74,7 +75,7 @@ import org.ops4j.pax.exam.spi.reactors.PerMethod;
  */
 @RunWith( RoboconfPaxRunner.class )
 @ExamReactorStrategy( PerMethod.class )
-public class AgentInitializationWithScopedInstanceTest extends DmTest {
+public class AgentInitializationTest extends DmTest {
 
 	private static final String APP_LOCATION = "my.app.location";
 
@@ -87,7 +88,6 @@ public class AgentInitializationWithScopedInstanceTest extends DmTest {
 
 		// We need to specify the classes we need
 		// and that come from external modules.
-		probe.addTest( AbstractTest.class );
 		probe.addTest( DmTest.class );
 		probe.addTest( TestUtils.class );
 
@@ -107,7 +107,9 @@ public class AgentInitializationWithScopedInstanceTest extends DmTest {
 	@Override
 	@Configuration
 	public Option[] config() throws Exception {
-		List<Option> options = getBaseOptions();
+
+		List<Option> options = new ArrayList<> ();
+		options.addAll( Arrays.asList( super.config()));
 
 		// Store the application's location
 		File resourcesDirectory = TestUtils.findApplicationDirectory( "lamp" );
@@ -115,23 +117,23 @@ public class AgentInitializationWithScopedInstanceTest extends DmTest {
 		options.add( systemProperty( APP_LOCATION ).value( appLocation ));
 
 		// Deploy the agent's bundles
+		String roboconfVersion = ItUtils.findRoboconfVersion();
 		options.add( mavenBundle()
 				.groupId( "net.roboconf" )
 				.artifactId( "roboconf-plugin-api" )
-				.version( getRoboconfVersion())
+				.version( roboconfVersion )
 				.start());
 
 		options.add( mavenBundle()
 				.groupId( "net.roboconf" )
 				.artifactId( "roboconf-agent" )
-				.version( getRoboconfVersion())
+				.version( roboconfVersion )
 				.start());
 
 		return options.toArray( new Option[ options.size()]);
 	}
 
 
-	@Override
 	@Test
 	public void run() throws Exception {
 
@@ -153,16 +155,14 @@ public class AgentInitializationWithScopedInstanceTest extends DmTest {
 		// There is no agent yet (no root instance was deployed)
 		Assert.assertEquals( 0, myResolver.handler.agentIdToAgent.size());
 
-		// Instantiate a new scoped instance
-		Instance scopedInstance = InstanceHelpers.findInstanceByPath( ma.getApplication(), "/Tomcat VM 1/Tomcat" );
-		Assert.assertNotNull( scopedInstance );
-		Assert.assertEquals( InstanceStatus.NOT_DEPLOYED, scopedInstance.getStatus());
+		// Instantiate a new root instance
+		Instance rootInstance = InstanceHelpers.findInstanceByPath( ma.getApplication(), "/MySQL VM" );
+		Assert.assertNotNull( rootInstance );
+		Assert.assertEquals( InstanceStatus.NOT_DEPLOYED, rootInstance.getStatus());
 
-		scopedInstance.getComponent().installerName( Constants.TARGET_INSTALLER );
-
-		this.manager.changeInstanceState( ma, scopedInstance, InstanceStatus.DEPLOYED_STARTED );
+		this.manager.changeInstanceState( ma, rootInstance, InstanceStatus.DEPLOYED_STARTED );
 		Thread.sleep( 800 );
-		Assert.assertEquals( InstanceStatus.DEPLOYED_STARTED, scopedInstance.getStatus());
+		Assert.assertEquals( InstanceStatus.DEPLOYED_STARTED, rootInstance.getStatus());
 
 		// A new agent must have been created
 		Assert.assertEquals( 1, myResolver.handler.agentIdToAgent.size());
@@ -170,33 +170,31 @@ public class AgentInitializationWithScopedInstanceTest extends DmTest {
 		Thread.sleep( 1000 );
 		Assert.assertFalse( agent.needsModel());
 		Assert.assertNotNull( agent.getScopedInstance());
-		Assert.assertEquals( "Tomcat", agent.getScopedInstance().getName());
-		Assert.assertEquals( 1, InstanceHelpers.buildHierarchicalList( agent.getScopedInstance()).size());
+		Assert.assertEquals( "MySQL VM", agent.getScopedInstance().getName());
+		Assert.assertEquals( 2, InstanceHelpers.buildHierarchicalList( agent.getScopedInstance()).size());
 
 		// Try to instantiate another VM
-		Instance anotherScopedInstance = InstanceHelpers.findInstanceByPath( ma.getApplication(), "/MySQL VM/MySQL" );
-		Assert.assertNotNull( anotherScopedInstance );
-		Assert.assertEquals( InstanceStatus.NOT_DEPLOYED, anotherScopedInstance.getStatus());
+		Instance anotherRootInstance = InstanceHelpers.findInstanceByPath( ma.getApplication(), "/Tomcat VM 1" );
+		Assert.assertNotNull( anotherRootInstance );
+		Assert.assertEquals( InstanceStatus.NOT_DEPLOYED, anotherRootInstance.getStatus());
 
-		anotherScopedInstance.getComponent().installerName( Constants.TARGET_INSTALLER );
-
-		this.manager.changeInstanceState( ma, anotherScopedInstance, InstanceStatus.DEPLOYED_STARTED );
+		this.manager.changeInstanceState( ma, anotherRootInstance, InstanceStatus.DEPLOYED_STARTED );
 		Thread.sleep( 800 );
-		Assert.assertEquals( InstanceStatus.DEPLOYED_STARTED, anotherScopedInstance.getStatus());
+		Assert.assertEquals( InstanceStatus.DEPLOYED_STARTED, anotherRootInstance.getStatus());
 
 		Assert.assertEquals( 2, myResolver.handler.agentIdToAgent.size());
-		Agent anotherAgent = myResolver.handler.agentIdToAgent.get( "/MySQL VM/MySQL @ test" );
+		Agent anotherAgent = myResolver.handler.agentIdToAgent.get( "/Tomcat VM 1 @ test" );
 		Assert.assertNotNull( anotherAgent );
 
 		Thread.sleep( 1000 );
 		Assert.assertFalse( anotherAgent.needsModel());
 		Assert.assertNotNull( anotherAgent.getScopedInstance());
-		Assert.assertEquals( "MySQL", anotherAgent.getScopedInstance().getName());
-		Assert.assertEquals( 1, InstanceHelpers.buildHierarchicalList( anotherAgent.getScopedInstance()).size());
+		Assert.assertEquals( "Tomcat VM 1", anotherAgent.getScopedInstance().getName());
+		Assert.assertEquals( 2, InstanceHelpers.buildHierarchicalList( anotherAgent.getScopedInstance()).size());
 
 		// Undeploy them all
-		this.manager.changeInstanceState( ma, scopedInstance, InstanceStatus.NOT_DEPLOYED );
-		this.manager.changeInstanceState( ma, anotherScopedInstance, InstanceStatus.NOT_DEPLOYED );
+		this.manager.changeInstanceState( ma, rootInstance, InstanceStatus.NOT_DEPLOYED );
+		this.manager.changeInstanceState( ma, anotherRootInstance, InstanceStatus.NOT_DEPLOYED );
 		Thread.sleep( 300 );
 		Assert.assertEquals( 0, myResolver.handler.agentIdToAgent.size());
 	}
