+++ miscellaneous/roboconf-integration-tests/src/test/java/net/roboconf/integration/tests/readme.txt
@@ -1,0 +1,3 @@
+* internal: utilities and so on.
+* paxrunner: tests whose code will run INSIDE Karaf. Useful to verify internal states and assertions.
+* servermode: tests that start a Karaf distribution and execute tests OUTSIDE the container.
